KVM: selftests: aarch64: Add helper function for the vpmu vcpu creation
Create a vcpu with vpmu would be a common requirement for the vpmu test,
so add the helper function for the vpmu vcpu creation. And use those
helper function in the vpmu_counter_access.c test.

Use this chance to delete the meaningless ASSERT about the pmuver,
because KVM does not advertise an IMP_DEF PMU to guests.

No functional changes intended.

Reviewed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Shaoqin Huang <shahuang@redhat.com>
Reviewed-by: Raghavendra Rao Ananta <rananta@google.com>
Message-Id: <20240619083200.1047073-2-shahuang@redhat.com>