KVM: arm64: vgic-its: Clear ite when discard frees an ite
When discard frees an ite, we do not invalidate the
corresponding ite. In the scenario of continuous save and
restore, there may be a situation where an ite is not saved
but is restored. This is unreasonable and may cause restore
to fail. This patch clears the corresponding ite when discard
frees a ite.

Link: https://lore.kernel.org/kvmarm/8f9d74fc-f9d9-43ac-a387-91ff804cfaf1@redhat.com/
Signed-off-by: Kunkun Jiang <jiangkunkun@huawei.com>
Message-Id: <20240620130650.1279-4-jiangkunkun@huawei.com>