KVM: arm64: Fix FFR offset calculation for pKVM host state save and restore
When saving and restoring the SVE state for the host we configure the
hardware for the maximum VL it supports, but when calculating offsets in
memory we use the maximum usable VL for the host. Since these two values
may not be the same this may result in data corruption in the case where
the PE supports a VL larger than the maximum usable VL for the host.  We
can just read the current VL from the hardware with an instruction so
use that instead of a saved value, we need to correct the value used to
lay out the stored data and this makes it clear that the layout is
consistent with the hardware configuration.

Fixes: b5b9955617bc ("KVM: arm64: Eagerly restore host fpsimd/sve state in pKVM")
Signed-off-by: Mark Brown <broonie@kernel.org>
Message-Id: <20240708-kvm-arm64-fix-pkvm-sve-vl-v5-3-d2175738456b@kernel.org>