KVM: arm64: Fix outdated comment about incomplete FGT bitmask definitions
Comments in the code say that our definitions for the FGT registers are
not the generated ones and are updated only as far as DDI0487 J.a.  In
commit 9ff67dd26a9e ("KVM: arm64: Use generated FGT RES0 bits instead of
specifying them") we updated to use the generated definitions.  As of
the release of DDI0487 K.a the XML in DD0601 2024-06 is in sync with the
ARM.  Remove the outdated comment about divergence.

Signed-off-by: Mark Brown <broonie@kernel.org>
Message-Id: <20240802-kvm-arm64-fgt-doc-v2-1-b90183df14f6@kernel.org>