KVM: selftests: arm64: Use generated defines for named system registers
Currently the get-reg-list test uses directly specified numeric values to
define system registers to validate. Since we already have a macro which
allows us to use the generated system register definitions from the main
kernel easily let's update all the registers where we have specified the
name in a comment to just use that macro. This reduces the number of
places where we need to validate the name to number mapping.

This conversion was done with the sed command:

  sed -i -E 's-ARM64_SYS_REG.*/\* (.*) \*/-KVM_ARM64_SYS_REG(SYS_\1),-' tools/testing/selftests/kvm/aarch64/get-reg-list.c

We still have a number of numerically specified registers, some of these
are reserved registers without defined names (eg, unallocated ID registers)
and others don't have kernel macro definitions yet.

No change in the generated output.

Suggested-by: Marc Zyngier <maz@kernel.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
Message-Id: <20240802-kvm-arm64-get-reg-list-v1-2-3a5bf8f80765@kernel.org>