arm64: Modify callback matches() fn to take a target info
In preparation of identifying the errata associated with a particular
target CPU, modify the matches fn to take a target pointer. For
capabilities representing system features, this is not used and
named as __unused. In subsequent patches, errata workarounds matches
function will make use of  the target pointer instead of the
read_cpuid_id() to check the cpu model.

No functional changes intended.

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>
Message-Id: <20241011075053.80540-2-shameerali.kolothum.thodi@huawei.com>