KVM: arm64: Add support for VMM to set migration target
Add a VM IOCTL to allow userspace(VMM) to set a number of migration
target CPUs. This can be used to tell KVM the list of targets this
VM will encounter in its lifetime. In subsequent patches, KVM will
use this information to enable errata associated with all the target
CPUs for this VM.

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>
Message-Id: <20241011075053.80540-3-shameerali.kolothum.thodi@huawei.com>