KVM: arm64: Introduce a helper to retrieve errata
And update the errata matches functions to use the target CPU values if
it is set.

Also introduce a "migration_safe_cap" to capabilities structure.
This should be a statically allocated constant for any migration safe errata.
This is because the existing "capability" value is a generated one and may
be renumbered and reordered, hence cannot be used to set the bits in
migration errata bitmap.

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>
Message-Id: <20241011075053.80540-4-shameerali.kolothum.thodi@huawei.com>