KVM: arm64: Add hypercall support for retrieving migration errata bitmap
If the VM requires migration and has set the list of target CPUs, retrieve
the errata bitmap for those targets. The Guest can use this hypercall
to get the errata bitmap. The bitmaps will be returned in R0-R3 registers.

ToDo: For now, reused one of the reserved pKVM hypercall value for this
as I noticed that the vendor_hyp_bmap size needs to be adjusted to have
anything > 63 set.

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>
Message-Id: <20241011075053.80540-5-shameerali.kolothum.thodi@huawei.com>