kselftest/arm64: Add 2024 dpISA extensions to hwcap test
We don't actually test SIGILL generation for CMPBR since the need to
branch makes it a pain to generate and the SIGILL detection would be
unreliable anyway. Since this should be very unusual we provide a stub
function rather than supporting a missing function.

The sigill functions aren't well sorted in the file so the ordering is a
bit random.

Signed-off-by: Mark Brown <broonie@kernel.org>
Message-Id: <20241028-arm64-2024-dpisa-v1-9-a38d08b008a8@kernel.org>