tee: optee: Add OP-TEE Mediator
The OP-TEE mediator is a software entity resposible for bridging
the gap between a KVM guest and OP-TEE in the secure world.

The guest's OP-TEE driver issues an SMC instruction after populating
its CPU registers with appropriate arguments. This SMC is trapped by
the hypervisor and control is passed to the OP-TEE mediator with the
vCPU state.

The mediator is resposible for manipulating the vCPU state accordingly
and keeping track of active transactions between the guest and the
TEE.

This implementation adds event handlers that gets hooked to the
TEE Mediator layer and are called when events such as guest
creation/destruction, mediator status check and guest SMC trap happen.

Important routines implemented:

	- optee_mediator_{create|destroy}_vm(): Sends an SMC to OP-TEE
	  notifying KVM guest creation or destruction.

	- optee_mediator_{create|destroy}_host(): Sends an SMC to OP-TEE
	  notifying host OP-TEE driver initalization/release (OP-TEE treats
	  all NS-EL1 entities as a guest, and is not aware of host privilege).

	- optee_mediator_forward_smc(): Changes vCPU register state as required
	  by OP-TEE and keeps track of standard calls and memory shared by guest.

The OP-TEE mediator is implemented in such a way that, the guest/VMM can remain
unmodified. The guest interacts with OP-TEE just as it would if it was running
natively.

Signed-off-by: Yuvraj Sakshith <yuvraj.kernel@gmail.com>
Message-Id: <20250401170527.344092-7-yuvraj.kernel@gmail.com>