KVM: arm64: Rename pkvm.enabled to pkvm.is_protected
The 'pkvm.enabled' field in struct kvm_protected_vm is confusingly named.
Its purpose is to indicate whether a VM is a pKVM protected VM, not
whether the VM itself is enabled or running.

For a non-protected VM, the VM can be fully active, yet this field would
be false. This ambiguity can lead to incorrect assumptions about the VM's
operational state and makes the code harder to reason about.

Rename the field to 'is_protected' to make it unambiguous that the flag
tracks the protected status of the VM.

No functional change intended.

Signed-off-by: Fuad Tabba <tabba@google.com>
Reviewed-by: Kunwu Chan <chentao@kylinos.cn>
Message-Id: <20250729120014.2799359-2-tabba@google.com>