arm64/fpsimd: Check enable bit for FA64 when saving EFI state
Currently when deciding if we need to save FFR when in streaming mode prior
to EFI calls we check if FA64 is supported by the system. Since KVM guest
support will mean that FA64 might be enabled and disabled at runtime switch
to checking if traps for FA64 are enabled in SMCR_EL1 instead.

Signed-off-by: Mark Brown <broonie@kernel.org>
Message-Id: <20250822-kvm-arm64-sme-v7-4-7a65d82b8b10@kernel.org>